scripts/requirements.txt
google-genai
youtube-transcript-api
httpx
beautifulsoup4
python-dotenv
lxml
